refactor: rename cloud_init to container_init, simplify resolver, update version policy and milestones
Agent-Logs-Url: https://github.com/peteralieber/yetaos/sessions/7eea36d6-81f9-47b3-9e41-d007404b0f06
