+++ lxc/cloud-init/tool/python.sh
@@ -1,9 +1,6 @@
-# pinned: python 3.11 from deadsnakes, uv 0.6.10
+# default: latest python3 from Ubuntu repos; version pinned at container creation time
 set -e
 export DEBIAN_FRONTEND=noninteractive
 apt-get update
-apt-get install -y software-properties-common
-add-apt-repository -y ppa:deadsnakes/ppa
-apt-get update
-apt-get install -y python3.11 python3.11-venv python3.11-dev python3-pip
-python3 -m pip install --no-cache-dir --upgrade uv==0.6.10
+apt-get install -y python3 python3-venv python3-dev python3-pip
+python3 -m pip install --no-cache-dir --upgrade uv
